Introduced bug fixed

+++ lib-device/include/max7219.h
@@ -2,7 +2,7 @@
  * @file max7219.h
  *
  */
-/* Copyright (C) 2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2020-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -77,7 +77,7 @@ public:
 	}
 
 	void WriteRegister(uint32_t nRegister, uint32_t nData, const bool bSpiSetup = true) {
-		const auto nSpiData = static_cast<uint8_t>(((nRegister & 0xFF) << 8) | (nData & 0xFF));
+		const auto nSpiData = static_cast<uint16_t>(((nRegister & 0xFF) << 8) | (nData & 0xFF));
 		Write(nSpiData, bSpiSetup);
 	}
 };
